スイッチ扉の追加

+++ Player.cpp
@@ -36,7 +36,9 @@ void Player::Draw(const int& offsetX, const int& offsetY)
 {
     draw->Draw(
         object,
-        pos + Vector3(offsetX, offsetY, 0.0f),
+        pos + Vector3(static_cast<float>(offsetX),
+                      static_cast<float>(offsetY),
+                      0.0f),
         Math::rotateZ(direction * Math::DEGREE_F * 90.0f),
         { 0.5f, 1.0f, 1.0f },
         DirectX::XMFLOAT4(1.0f, 1.0f, 1.0f, 1.0f),
